# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Real Estate AI Assistant API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List property listings
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        price:
          type: number
          nullable: true
        beds:
          type: integer
          nullable: true
        baths:
          type: number
          nullable: true
        sqft:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, pending, sold, off_market]
        notes:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, idx, mls, csv]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - address
        - status
        - source
        - createdAt
        - updatedAt
    CreateListingBody:
      type: object
      properties:
        address:
          type: string
        price:
          type: number
        beds:
          type: integer
        baths:
          type: number
        sqft:
          type: integer
        status:
          type: string
          enum: [active, pending, sold, off_market]
        notes:
          type: string
        source:
          type: string
          enum: [manual, idx, mls, csv]
      required:
        - address
        - status
        - source
...